Updated the Date strings in activities to use a timestamp.
To allow for a users locale to be different from the DT instance locale, we need to get a timestamp from the server and compile the date string based on the users locale. To do this in the main DT theme I made some changes in https://github.com/DiscipleTools/disciple-tools-theme/pull/925 to change the strings we save for Baptism date and Group start/end dates to include a time stamp wrapped in curly brackets. Then in JS locally we replace the timestamp with a formatted date string in the users locale.